requirements.txt
fastapi
uvicorn
pydantic
google-cloud-aiplatform
jinja2